### 2016-11-14 #115 @whedon
[PRE REVIEW]: RAILS and Cobbler: Scaffolding and automated finishing of draft genomes using long sequences
**Submitting author:** @warrenlr (<a href="http://orcid.org/0000-0002-9890-2293">Rene L. Warren</a>)
**Repository:** <a href="https://github.com/warrenlr/RAILS" target ="_blank">https://github.com/warrenlr/RAILS</a>
**Version:** v1.1
