### 2016-11-14 #116 @whedon
[REVIEW]: RAILS and Cobbler: Scaffolding and automated finishing of draft genomes using long sequences
**Submitting author:** @warrenlr (<a href="http://orcid.org/0000-0002-9890-2293">Rene L. Warren</a>)
**Repository:** <a href="https://github.com/warrenlr/RAILS" target ="_blank">https://github.com/warrenlr/RAILS</a>
**Version:** v1.1
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.167092" target="_blank">https://dx.doi.org/10.5281/zenodo.167092</a>
