### 2016-11-15 3576b74
fixed markdown for tables, titles and added community guidelines to readme
