+++ readme.md
@@ -1,15 +1,15 @@
-RAILS v1.1 and Cobbler v0.2 Rene L. Warren, 2014-2016
+### RAILS v1.1 and Cobbler v0.2 Rene L. Warren, 2014-2016
 email: rwarren at bcgsc.ca
 
-Name
-----
+### Name
+### ----
 
 RAILS: Radial Assembly Improvement by Long Sequence Scaffolding
 Cobbler: Gap-filling with long sequences
 
 
-Description
------------
+### Description
+### -----------
 
 RAILS and Cobbler are genomics application for scaffolding and automated finishing of genome assemblies with long DNA sequences.
 They can be used to scaffold & finish high-quality draft genome assemblies with any long, preferably high-quality, sequences such as scaftigs/contigs from another genome draft. 
@@ -23,32 +23,41 @@ RAILS is an all-in-one scaffolder and gap-filler. Its process is similar to that
 
 You can test the software by executing "runme.sh" in the test folder. A simulated SARS genome assembly is provided to test the software. 
 
-Implementation and requirements
--------------------------------
+### Implementation and requirements
+### -------------------------------
 
 RAILS and Cobbler are implemented in PERL and run on any OS where PERL is installed.
 
 
-Install
--------
+### Community guidelines:
+### ---------------------
+
+I encourage the community to contribute to the development of this software, by providing suggestions for improving the code and/or directly contributing to the open source code for these tools. Users and developers may report software issues, bug fix requests, comments, etc, at <https://github.com/warrenlr/RAILS>
+
+
+### Install
+### -------
 
 Download the tar ball, gunzip and extract the files on your system using:
 
 gunzip rails_v1-1.tar.gz
 tar -xvf rails_v1-1.tar
 
+Alternatively, individual tools are available within the github repository
 
-Dependencies
-------------
+
+### Dependencies
+### ------------
 
 Make sure you have installed bwa (Version: 0.7.15-r1140) and that is is in your path.
 
 
-Test data
----------
+### Test data
+### ---------
 
 Go to ./test
 (cd test)
+
 1. SARS
 execute runme.sh
 (./runme.sh)
@@ -58,8 +67,8 @@ execute runmeHuman.sh (will take a while to run)
 (./runmeHuman.sh)
 
 
-Usage
------
+### Usage
+### -----
 
 ./runRAILS.sh
 Usage: runRAILS.sh <FASTA assembly .fa> <FASTA long sequences .fa> <anchoring sequence length eg. 250> <min sequence identity 0.95>
@@ -96,8 +105,8 @@ Usage: ./RAILS [v1.1]
 -v  Runs in verbose mode (-v 1 = yes, default = no, optional)
 
 
-How it works
-------------
+### How it works
+### ------------
 
 The pipeline is detailed in the provided script runRAILS.sh
 
@@ -125,45 +134,46 @@ Warren RL, Yang C, Vandervalk BP, Behsaz B, Lagman A, Jones SJ, Birol I. 2015. L
 Warren RL, Sutton GG, Jones SJM, Holt RA.  2007.  Assembling millions of short DNA sequences using SSAKE.  Bioinformatics. 23(4):500-501. DOI: 10.1093/bioinformatics/btl629
 
 
-Runs on the human genome
-------------------------
+### Runs on the human genome
+### ------------------------
 
[…]
