+++ readme.md
@@ -1,15 +1,13 @@
-### RAILS v1.1 and Cobbler v0.2 Rene L. Warren, 2014-2016
+# RAILS v1.1 and Cobbler v0.2 Rene L. Warren, 2014-2016
 email: rwarren at bcgsc.ca
 
 ### Name
-### ----
 
 RAILS: Radial Assembly Improvement by Long Sequence Scaffolding
 Cobbler: Gap-filling with long sequences
 
 
 ### Description
-### -----------
 
 RAILS and Cobbler are genomics application for scaffolding and automated finishing of genome assemblies with long DNA sequences.
 They can be used to scaffold & finish high-quality draft genome assemblies with any long, preferably high-quality, sequences such as scaftigs/contigs from another genome draft. 
@@ -24,19 +22,16 @@ RAILS is an all-in-one scaffolder and gap-filler. Its process is similar to that
 You can test the software by executing "runme.sh" in the test folder. A simulated SARS genome assembly is provided to test the software. 
 
 ### Implementation and requirements
-### -------------------------------
 
 RAILS and Cobbler are implemented in PERL and run on any OS where PERL is installed.
 
 
 ### Community guidelines:
-### ---------------------
 
 I encourage the community to contribute to the development of this software, by providing suggestions for improving the code and/or directly contributing to the open source code for these tools. Users and developers may report software issues, bug fix requests, comments, etc, at <https://github.com/warrenlr/RAILS>
 
 
 ### Install
-### -------
 
 Download the tar ball, gunzip and extract the files on your system using:
 
@@ -47,13 +42,11 @@ Alternatively, individual tools are available within the github repository
 
 
 ### Dependencies
-### ------------
 
 Make sure you have installed bwa (Version: 0.7.15-r1140) and that is is in your path.
 
 
 ### Test data
-### ---------
 
 Go to ./test
 (cd test)
@@ -68,7 +61,6 @@ execute runmeHuman.sh (will take a while to run)
 
 
 ### Usage
-### -----
 
 ./runRAILS.sh
 Usage: runRAILS.sh <FASTA assembly .fa> <FASTA long sequences .fa> <anchoring sequence length eg. 250> <min sequence identity 0.95>
@@ -106,7 +98,6 @@ Usage: ./RAILS [v1.1]
 
 
 ### How it works
-### ------------
 
 The pipeline is detailed in the provided script runRAILS.sh
 
@@ -135,45 +126,43 @@ Warren RL, Sutton GG, Jones SJM, Holt RA.  2007.  Assembling millions of short D
 
 
 ### Runs on the human genome
-### ------------------------
 
 On a human draft assembly, cobbler patched over 65% of the gaps using 1, 2.5, 5, 15 kb long DNA sequences simulated from the human genome reference. The Pearson correlation between the predicted gap sizes and the size of patched gaps is R=0.8150
 
 
-Table 1. Patching gaps with Cobbler using simulated 1, 2.5, 5, 15kbp simulated long sequences from human genome reference GRCh38.
-|Total gaps | 148,091|
-|Number of gaps patched | 95,523|
-|Proportion of gaps patched | 65.1%|
-|Average length (bp) | 343.39|
-|Length st.dev +/- | 931.12|
-|Total bases added | 32,801,755|
-|Largest gap resolved (bp) | 13,662|
-|Shortest gap resolved (bp) | 1 |
[…]
