### 2016-11-15 fea334c
further edits to markdown
