### 2016-11-15 772f319
Added citations to paper,md, made minor edits to the text, added tables
