+++ paper/paper.bib
@@ -6,6 +6,23 @@
   urldate = {2016-11-10}
 }
 
+@Article{Pagani2012,
+author="Pagani, I.
+and Liolios, K.
+and Jansson, J.
+and Chen, I. -. M.
+and Smirnova, T.
+and Nosrat, B.
+and Markowitz, V. M.
+and Kyrpides, N. C.",
+title="The Genomes OnLine Database (GOLD) v. 4: status of genomic and metagenomic projects and their associated metadata",
+journal="Nucleic Acids Res",
+year="2012",
+volume="40",
+doi="10.1093/nar/gkr1100",
+url="http://dx.doi.org/10.1093/nar/gkr1100"
+}
+
 @Article{Paulino2015,
 author="Paulino, Daniel
 and Warren, Ren{\'e} L.
@@ -25,6 +42,36 @@ doi="10.1186/s12859-015-0663-4",
 url="http://dx.doi.org/10.1186/s12859-015-0663-4"
 }
 
+@Article{Tsai2010,
+author="Tsai, Isheng J.
+and Otto, Thomas D.
+and Berriman, Matthew",
+title="Improving draft assemblies by iterative mapping and assembly of short reads to eliminate gaps",
+journal="Genome Biology",
+year="2010",
+volume="11",
+number="4",
+pages="R41",
+abstract="Advances in sequencing technology allow genomes to be sequenced at vastly decreased costs. However, the assembled data frequently are highly fragmented with many gaps. We present a practical approach that uses Illumina sequences to improve draft genome assemblies by aligning sequences against contig ends and performing local assemblies to produce gap-spanning contigs. The continuity of a draft genome can thus be substantially improved, often without the need to generate new data.",
+issn="1474-760X",
+doi="10.1186/gb-2010-11-4-r41",
+url="http://dx.doi.org/10.1186/gb-2010-11-4-r41"
+}
+
+@Article{Boetzer2012,
+author="Boetzer, Marten
+and Pirovano, Walter",
+title="Toward almost closed genomes with GapFiller",
+journal="Genome Biology",
+year="2012",
+volume="13",
+number="6",
+pages="R56",
+abstract="De novo assembly is a commonly used application of next-generation sequencing experiments. The ultimate goal is to puzzle millions of reads into one complete genome, although draft assemblies usually result in a number of gapped scaffold sequences. In this paper we propose an automated strategy, called GapFiller, to reliably close gaps within scaffolds using paired reads. The method shows good results on both bacterial and eukaryotic datasets, allowing only few errors. As a consequence, the amount of additional wetlab work needed to close a genome is drastically reduced. The software is available at                   http://www.baseclear.com/bioinformatics-tools/                                  .",
+issn="1474-760X",
+doi="10.1186/gb-2012-13-6-r56",
+url="http://dx.doi.org/10.1186/gb-2012-13-6-r56"
+}
 
 @Article{Warren2015,
 author="Warren, Ren{\'e} L.
