### 2016-11-15 185e67c
minor edits
