### 2016-11-15 #116 @andrewjpage
Good work @warrenlr and thanks for making all the changes I suggested.
