### 2016-11-16 #116 @arfon
Thanks for completing the review @andrewjpage. I've reopened the issue as I haven't finished all of the editorial tasks yet for this paper.

@warrenlr - one small suggestion - would you mind renaming the `gpl-3.0.md` file to `LICENSE` - this is a more 'standard' way to include a license file in a repository.
