### 2016-11-16 #116 @arfon
Thanks @warrenlr. At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
