### 2016-11-16 d27a041
minor edits to keywords for paper.md
