### 2016-11-17 #116 @arfon
@biorelated - many thanks for editing this submission. @andrewjpage - thanks for the review!

@warrenlr - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00116 🎉 🚀 💥
